requirements.txt
discord.py
pillow
flask
pip install -r requirements.txt
